requirements.txt
wandb==0.17.8
weave==0.50.14
python-dotenv==1.0.1
datasets==2.21.0
openai==1.39.0